nx_kit: Fixed race in C-wrapper unit test; minor code improvements
<==== Mercurial data ====>
Branch: vms

+++ open/artifacts/nx_kit/src/nx/kit/ini_config.cpp
@@ -289,7 +289,7 @@ bool Param<double>::reload(const std::string* value, std::ostream* output)
         if (errno == ERANGE || *pEnd != '\0')
             error = " [invalid value]";
         else
-            *pValue = (double) v;
+            *pValue = v;
     }
     printValueLine(output, *pValue, " = ", error, *pValue == defaultValue);
     return oldValue != *pValue;
